Update to Pete's fix branch
Pete's fix branch for ADC stuff is till working on
getting upstreamed. In the meantime, it has been rebased
on the latest Zephyr. Included here are the changes required

The latest tmp112 sensor driver tries to configure the device in its
init function. This poses a problem because we need to turn on the
3.3 power supply first. The fix here is to turn on essential
power supplies before the I2C driver init and sensor inits