Change liquid lens temp sensor to use regular data format
We were using the extended data format, but the problem with
changing the data format on this sensor is that it takes up to 8
seconds for the format change to take effect. In that time, we
can accidentally interpret the temperature as 2 times higher
than it really is. The solution is to not change the sensor's
default data format, whose precision is good enough for our
purposes anyway.