Fix NMEA partial message counter value
When we send a partial NMEA message via Protobuf, we include a counter
value that when even indicates part 1 of a 2-part partial NMEA message.
When the counter is odd, it indicates part 2 of a 2-part partial NMEA message.

This commit fixes a possible code path where we accidentally switch sending
even and odd values.