front unit: diamond b3: fuse circuit replaced with power switch (#18)
* front unit: support different signals on B3

Add support for different signals on Front Unit 6.2 which is used in
the Diamond B3 build. On Front Unit 6.2 the electronic fuse circuit
was removed and replaced by a simple power switch.

* front unit: bug fix

Initialize fuse_reset even if fuse was not blown. Otherwise the fuse_reset
signal is floating and B3 Front Units can't be used with POC2 Mainboards.