fix(pubsub): compute McuMessage wrapper size without throwing error
because we want to minimize memory used to store messages to send, we
compute the number of bytes taken to wrap the payload to send into an
McuMessage.
The number of bytes is guessed thanks to the McuMessage max size and all
 the payloads max sizes which are all known at compile time.
In case the biggest payload among the 4 available changes, the
computation needs to change: we now use the max of all 4 so that we
don't have to throw an error in case the biggest payload changes.

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>