feat(motion): validate that icm40627 motion sensor is responding
initialize icm40627 sensor by resetting it and checking
the WHO_AM_I register.

the driver is based on the icm42670 present in the zephyr repository
with a few changes: SPI -> I2C, changed WHO_AM_I response, added 1
int-gpios (still unused).
this driver is kept locally for now (no attempt to bring it upstream)
as it is still experimental and hasn't been tested.

Accelerometer values can be read and values are responding to motion,
but resolution might be wrong though (in m/s2):

`<inf> motion: Accel data for X: 1.606509 Y: 0.836773 Z: -3.-578134`

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>