fix(shutdown): proper jetson shutdown (#213)
power-enable pin is connected over a gpio expander, unusable from
the interrupt, so add task into work queue, which is gonna be handled
within the 100ms constraint given by hardware.

on the jetson side, the power-enable must be pulled low within 10us,
which is ensured by the electronic design, but the circuit pulls
power-enable low for at least 100ms (200ms measured), and so the mcu is
expected to permanently force the state low within these ~100ms.

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>