fix(boot): reliable sd card initialization with late stage super-caps (#244)
sometimes sd card isn't recognized on cold boot: no battery
plugged in for a long time, then with a battery freshly inserted
and instant boot, we can easily reproduce the issue.

the difference between cold boot, and a fresh reboot is mainly
into the voltage at the (super) capacitors.
The need for current when starting to charge the super-caps
might alter the voltage at the sd card and/or other components
that are not correctly initialized.
The super-caps are thus enabled later in the boot process
(after UEFI boots, which needs ~15 seconds).
With this new boot sequence, the sd card seems to be always
correctly detected.

---------

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>