chore(ir-leds): allow setting up ir-led & fps on unsafe distance (#374)
when an object is too close (with 1d-tof detection), setting the ir-led
on duration & fps was prevented and a "forbidden" ack message was
responded, leading to camera timeouts.
Prefer to allow setting the ir led time & fps so that the camera is
triggered (with no light though), but at least orb-core doesn't crash
and the 1d-tof might detect a valid range from the user in front in the
next seconds to allow the signup to proceed.

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>