fix(ir_camera_system): safety feature from 1d-tof distance
correctly disable ir leds based on proximity detection of the 1d-tof
when safety is a concern, on_time_in_us is reset to 0, which overrides
any previous command from the jetson.
when the jetson tries to set a new value, the safety condition is
checked and on_time_in_us is set only when it's a safe operation.

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>