fix(tof_1d): reconfigure sensor (#388)
the sensor was reconfigured even if nothing in front of the orb, which
isn't a case where it needs to be reconfigured; on top of that, in that
case, the acquisition needs to be stopped for the sensor to be
reconfigured. This use case shouldn't be needed, but safer to stop the
sensor first, then reconfigure.

this fix should allow reconfiguration after disconnection, and only
after disconnection, by pulling a config from the sensor on every new
loop to ensure correct connection to the sensor.

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>